+++ NEWS.md
@@ -8,6 +8,7 @@
 * `across()` handles data frames with 0 columns (#5523). 
 
 * `mutate()` always keeps grouping variables, unconditional to `.keep=` (#5582).
+* `mutate()` and friends preserves row names in data frames once more (#5418).
 
 # dplyr 1.0.2